FIX-#2442: added tests for base backend
