+++ .github/workflows/ci.yml
@@ -191,6 +191,8 @@ jobs:
       - shell: bash -l {0}
         run: python -m pytest modin/test/test_envvar_catcher.py
       - shell: bash -l {0}
+        run: python -m pytest modin/test/backends/base/test_internals.py
+      - shell: bash -l {0}
         run: python -m pytest modin/test/backends/pandas/test_internals.py
 
   test-defaults:
@@ -226,6 +228,8 @@ jobs:
           conda list
       - name: Install HDF5
         run: sudo apt update && sudo apt install -y libhdf5-dev
+      - shell: bash -l {0}
+        run: python -m pytest modin/test/backends/base/test_internals.py --backend=${{ matrix.backend }}
       - shell: bash -l {0}
         run: pytest modin/pandas/test/dataframe/test_binary.py --backend=${{ matrix.backend }}
       - shell: bash -l {0}
